Moved FoodItemDetailsView.java to screens
